Use `YAML` 6.1.0
