+++ plugin/src/main/java/net/dzikoysk/funnyguilds/config/message/MessageService.java
@@ -1,13 +1,19 @@
 package net.dzikoysk.funnyguilds.config.message;
 
+import dev.peri.yetanothermessageslibrary.BukkitMessageService;
 import dev.peri.yetanothermessageslibrary.SimpleSendableMessageService;
-import dev.peri.yetanothermessageslibrary.viewer.BukkitViewerService;
+import dev.peri.yetanothermessageslibrary.config.serdes.SerdesMessages;
+import dev.peri.yetanothermessageslibrary.viewer.BukkitViewerDataSupplier;
+import dev.peri.yetanothermessageslibrary.viewer.ViewerFactory;
+import eu.okaeri.configs.ConfigManager;
+import eu.okaeri.configs.yaml.bukkit.YamlBukkitConfigurer;
 import java.io.File;
 import java.io.IOException;
 import net.dzikoysk.funnyguilds.FunnyGuilds;
 import net.dzikoysk.funnyguilds.FunnyGuildsLogger;
-import net.dzikoysk.funnyguilds.config.ConfigurationFactory;
 import net.dzikoysk.funnyguilds.config.PluginConfiguration;
+import net.dzikoysk.funnyguilds.config.serdes.DecolorTransformer;
+import net.dzikoysk.funnyguilds.config.serdes.FunnyTimeFormatterTransformer;
 import net.dzikoysk.funnyguilds.shared.FunnyIOUtils;
 import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import org.bukkit.Bukkit;
@@ -21,7 +27,8 @@ public class MessageService extends SimpleSendableMessageService<CommandSender, 
 
     public MessageService(FunnyGuilds plugin, BukkitAudiences adventure) {
         super(
-                new BukkitViewerService(plugin, adventure),
+                new BukkitViewerDataSupplier(adventure),
+                ViewerFactory.create(BukkitMessageService.wrapScheduler(plugin)),
                 (viewerService, localeSupplier, messageSupplier) -> new FunnyMessageDispatcher(viewerService, localeSupplier, messageSupplier, user -> Bukkit.getPlayer(user.getUUID()))
         );
         this.adventure = adventure;
@@ -62,9 +69,24 @@ public class MessageService extends SimpleSendableMessageService<CommandSender, 
                     logger.warning("New language file will be created with default values");
                 }
             }
-            messageService.registerRepository(locale, ConfigurationFactory.createMessageConfiguration(localeFile));
+            messageService.registerRepository(locale, createMessageConfiguration(localeFile));
         });
         return messageService;
+    }
+
+    public static MessageConfiguration createMessageConfiguration(File messageConfigurationFile) {
+        return ConfigManager.create(MessageConfiguration.class, (it) -> {
+            it.withConfigurer(new YamlBukkitConfigurer());
+            it.withSerdesPack(registry -> {
+                registry.register(new DecolorTransformer());
+                registry.register(new FunnyTimeFormatterTransformer());
+                registry.register(new SerdesMessages());
+            });
+
+            it.withBindFile(messageConfigurationFile);
+            it.saveDefaults();
+            it.load(true);
+        });
     }
 
 }
